[MOB-4876] SwiftUI/UIKit 컴포넌트 네이밍 컨벤션 적용
ComponentsKit 등 타 디자인 시스템 사례를 참고해 두 프레임워크 구현을 구분하는 네이밍 규칙을 도입.

- SwiftUI: `SU` prefix 추가 (BezierButton → SUBezierButton)
- UIKit: prefix/suffix 없이 기본 이름 (BezierButtonUIView → BezierButton)
- 공유 spec 타입(Size/Variant/Semantic/Resizing)은 그대로 유지
